CollectionExtension.swift was added to handle additional functionality of collections
