+++ src/PrebidMobile/PrebidMobile/CollectionExtension.swift
@@ -1,0 +1,27 @@
+//
+//  CollectionExtension.swift
+//  PrebidMobile
+//
+//  Created by Alex on 7/15/19.
+//  Copyright © 2019 AppNexus. All rights reserved.
+//
+
+import Foundation
+
+extension Array {
+    public func toDictionary<Key: Hashable>(with selectKey: (Element) -> Key) -> [Key: Element] {
+        var dict = [Key: Element]()
+        for element in self {
+            dict[selectKey(element)] = element
+        }
+        return dict
+    }
+}
+
+extension Dictionary {
+    mutating func merge(dict: [Key: Value]) {
+        for (k, v) in dict {
+            updateValue(v, forKey: k)
+        }
+    }
+}
