+++ src/PrebidMobile/PrebidMobile/Constants.swift
@@ -97,21 +97,3 @@ extension UIDevice {
     }
 
 }
-
-extension Dictionary {
-    mutating func merge(dict: [Key: Value]) {
-        for (k, v) in dict {
-            updateValue(v, forKey: k)
-        }
-    }
-}
-
-extension Array {
-    public func toDictionary<Key: Hashable>(with selectKey: (Element) -> Key) -> [Key: Element] {
-        var dict = [Key: Element]()
-        for element in self {
-            dict[selectKey(element)] = element
-        }
-        return dict
-    }
-}
